# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "APIs For Thrift Store",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/message/store": {
            "post": {
                "tags": [
                    "Messages"
                ],
                "summary": "Store a new message",
                "description": "Stores a new message in the database",
                "operationId": "storeMessage",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Message"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Message"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                },
                                "example": "Invalid data provided"
                            }
                        }
                    }
                }
            }
        },
        "/api/message/show/{id}": {
            "get": {
                "tags": [
                    "Messages"
                ],
                "summary": "Get a message by ID",
                "description": "Retrieves a message by its ID",
                "operationId": "getMessageById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the Message",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Message"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                },
                                "example": "Message not found"
                            }
                        }
                    }
                }
            }
        },
...
        "/api/Notification/ShowNotificationsByReceptientId/{ReceptientId}": {
            "get": {
                "tags": [
                    "Notifications"
                ],
                "summary": "Get notifications by Recipient ID",
                "description": "Retrieves notifications associated with a specific Recipient by their ID",
                "operationId": "getNotificationsByReceptientId",
                "parameters": [
                    {
                        "name": "ReceptientId",
                        "in": "path",
                        "description": "ID of the Recipient",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Notification"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/Notification/show/{id}": {
            "get": {
                "tags": [
                    "Notifications"
                ],
                "summary": "Get a notification by ID",
                "description": "Retrieves a notification by its ID",
                "operationId": "getNotificationById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the Notification",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Notification"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                },
                                "example": "Notification not found"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Message": {
                "title": "Message",
                "description": "Message model",
                "properties": {
                    "dateM": {
                        "description": "Date of the message",
                        "type": "string",
                        "format": "date"
                    },
                    "description": {
                        "description": "Description of the message",
                        "type": "string"
                    },
                    "sender_id": {
                        "description": "ID of the sender user",
                        "type": "integer"
                    },
                    "recipient_id": {
                        "description": "ID of the recipient user",
                        "type": "integer"
                    }
                },
                "type": "object"
            },
            "Notification": {
                "title": "Notification",
                "description": "Notification model",
                "properties": {
                    "message_id": {
                        "description": "ID of the message associated with the notification",
                        "type": "integer"
                    },
                    "message_description": {
                        "description": "Description of the message",
                        "type": "string"
                    },
                    "sender_msg": {
                        "description": "Sender of the message",
                        "type": "string"
                    },
                    "receptient_msg": {
                        "description": "Recipient of the message",
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}